Add static_cast to rect_2d initializer lists
The C++11 compiler does not narrow non constant expressions in the
initializer lists.

+++ timestep/timestep_view.cpp
@@ -38,7 +38,12 @@ static void image_view_render(timestep_view *v, context_2d *ctx) {
     if (map && map->url) {
         float scale_x = (float)v->width / (map->margin_left + map->width + map->margin_right);
         float scale_y = (float)v->height / (map->margin_top + map->height + map->margin_bottom);
-        rect_2d src_rect = {map->x, map->y, map->width, map->height};
+        rect_2d src_rect = {
+          static_cast<float>(map->x),
+          static_cast<float>(map->y),
+          static_cast<float>(map->width),
+          static_cast<float>(map->height)
+        };
         rect_2d dest_rect = {
             scale_x * map->margin_left,
             scale_y * map->margin_top,
@@ -179,7 +184,7 @@ void timestep_view_wrap_render(timestep_view *v, context_2d *ctx, JS_OBJECT_WRAP
     context_2d_translate(ctx, -v->anchor_x, -v->anchor_y);
 
     if (v->clip) {
-        rect_2d r = {0, 0, v->width, v->height};
+        rect_2d r = {0, 0, static_cast<float>(v->width), static_cast<float>(v->height)};
         context_2d_setClip(ctx, r);
     }
 
@@ -196,7 +201,7 @@ void timestep_view_wrap_render(timestep_view *v, context_2d *ctx, JS_OBJECT_WRAP
     if (v->background_color.a > 0) {
         //LOG("render %i with background %f %f %f %f", v->uid, v->background_color.r, v->background_color.g, v->background_color.b, v->background_color.a);
 
-        rect_2d r = {0, 0, v->width, v->height};
+        rect_2d r = {0, 0, static_cast<float>(v->width), static_cast<float>(v->height)};
         context_2d_fillRect(ctx, &r, &v->background_color);
     }
